24-July-2025/quiz-me/requirements.txt
# Core Framework Dependencies
gradio>=4.0.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0

# LLM Provider
openai>=1.0.0

# Data Models and Validation
pydantic>=2.0.0
typing-extensions>=4.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.8.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.0.0

# Optional: Enhanced Features
redis>=4.0.0
streamlit>=1.28.0 